+++ src/Confluent.Kafka/Admin/OffsetSpec.cs
@@ -27,17 +27,34 @@ namespace Confluent.Kafka.Admin
         /// <summary>
         ///     EarliestSpec : OffsetSpec denotes the EarliestSpec corresponding to the earliest offset produced.
         /// </summary>
-        public class EarliestSpec : OffsetSpec { }
+        public class EarliestSpec : OffsetSpec
+        {
+            internal override long Value()
+            {
+                return -2;
+            }
+        }
 
         /// <summary>
         ///     LatestSpec : OffsetSpec denotes the LatestSpec corresponding to the latest offset produced.
         /// </summary>
-        public class LatestSpec : OffsetSpec { }
+        public class LatestSpec : OffsetSpec
+        {
+            internal override long Value()
+            {
+                return -1;
+            }
+        }
 
         /// <summary>
         ///     MaxTimestampSpec : OffsetSpec denotes the MaxTimestampSpec as timestamp can be set on the client side which might differ with the LatestSpec.
         /// </summary>
-        public class MaxTimestampSpec : OffsetSpec { }
+        public class MaxTimestampSpec : OffsetSpec {
+            internal override long Value()
+            {
+                return -3;
+            }
+        }
 
         /// <summary>
         ///     TimestampSpec : OffsetSpec denotes the TimestampSpec corresponding to a particular timestamp.
@@ -47,20 +64,24 @@ namespace Confluent.Kafka.Admin
             /// <summary>
             ///     Timestamp for the OffsetSpec.
             /// </summary>
-            public long Timestamp { get; set; }
+            public long Timestamp { get; }
 
             /// <summary>
             ///     Sets the Timestamp for the TimestampSpec.
             /// </summary>
-            public TimestampSpec(long timestamp)
+            internal TimestampSpec(long timestamp)
             {
                 Timestamp = timestamp;
             }
 
+            internal override long Value()
+            {
+                return Timestamp;
+            }
         }
 
         /// <summary>
-        /// Used to retrieve the latest offset of a partition.
+        ///     Used to retrieve the latest offset of a partition.
         /// </summary>
         public static OffsetSpec Latest()
         {
@@ -68,7 +89,7 @@ namespace Confluent.Kafka.Admin
         }
 
         /// <summary>
-        /// Used to retrieve the earliest offset of a partition.
+        ///     Used to retrieve the earliest offset of a partition.
         /// </summary>
         public static OffsetSpec Earliest()
         {
@@ -76,8 +97,8 @@ namespace Confluent.Kafka.Admin
         }
 
         /// <summary>
-        /// Used to retrieve the earliest offset whose timestamp is greater than
-        /// or equal to the given timestamp in the corresponding partition.
+        ///     Used to retrieve the earliest offset whose timestamp is greater than
+        ///     or equal to the given timestamp in the corresponding partition.
         /// </summary>
         /// <param name="timestamp">Timestamp in milliseconds.</param>
         public static OffsetSpec ForTimestamp(long timestamp)
@@ -86,11 +107,13 @@ namespace Confluent.Kafka.Admin
         }
 
         /// <summary>
-        /// Used to retrieve the offset with the largest timestamp of a partition.
+        ///     Used to retrieve the offset with the largest timestamp of a partition.
         /// </summary>
         public static OffsetSpec MaxTimestamp()
         {
             return new MaxTimestampSpec();
         }
-}
+        
+        internal abstract long Value();
+    }
 }