+++ test/Confluent.Kafka.IntegrationTests/Tests/AdminClient_ListOffsets.cs
@@ -28,67 +28,50 @@ namespace Confluent.Kafka.IntegrationTests
         public async void AdminClient_ListOffsets(string bootstrapServers)
         {
             LogToFile("start AdminClient_ListOffsets");
-            using var topic = new TemporaryTopic(bootstrapServers, 1);
             
-            using (var producer = new ProducerBuilder<Null, string>(new ProducerConfig { BootstrapServers = bootstrapServers }).Build())
-            {   
-                long basetimestamp = 10000000;
-                await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 100, TimestampType.CreateTime)});
-                await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 400, TimestampType.CreateTime)});
-                await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 250, TimestampType.CreateTime)});
-                producer.Flush(new TimeSpan(0, 0, 10));
-            }
+            using var topic = new TemporaryTopic(bootstrapServers, 1);
+            using var producer = new ProducerBuilder<Null, string>(new ProducerConfig { BootstrapServers = bootstrapServers }).Build();
+            using var adminClient = new AdminClientBuilder(new AdminClientConfig { BootstrapServers = bootstrapServers }).Build();
+            
+            long basetimestamp = 10000000;
+            await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 100, TimestampType.CreateTime)});
+            await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 400, TimestampType.CreateTime)});
+            await producer.ProduceAsync(topic.Name, new Message<Null, string> { Value = "Producer Message", Timestamp = new Timestamp(basetimestamp + 250, TimestampType.CreateTime)});
+            producer.Flush(new TimeSpan(0, 0, 10));
+            
             var timeout = TimeSpan.FromSeconds(30);
-            ListOffsetsOptions options = new ListOffsetsOptions(){RequestTimeout = timeout, IsolationLevel = Confluent.Kafka.Admin.IsolationLevel.ReadUncommitted};
+            ListOffsetsOptions options = new ListOffsetsOptions(){RequestTimeout = timeout, IsolationLevel = IsolationLevel.ReadUncommitted};
 
-            using (var adminClient = new AdminClientBuilder(new AdminClientConfig { BootstrapServers = bootstrapServers }).Build())
+            var testFixtures = new List<Tuple<OffsetSpec, Offset>>
             {
-                var topicPartitionOffsetSpecs = new List<TopicPartitionOffsetSpec>();
-                topicPartitionOffsetSpecs.Add(new TopicPartitionOffsetSpec {
-                    TopicPartition = new TopicPartition(topic.Name, new Partition(0)),
-                    OffsetSpec = OffsetSpec.Earliest()
-                });
-                var ListOffsetsResult = await adminClient.ListOffsetsAsync(topicPartitionOffsetSpecs,options);
-                foreach(var ListOffsetsResultInfo in ListOffsetsResult.ListOffsetsResultInfos)
+                Tuple.Create(OffsetSpec.Earliest(), new Offset(0)),
+                Tuple.Create(OffsetSpec.Latest(), new Offset(3)),
+                Tuple.Create(OffsetSpec.MaxTimestamp(), new Offset(1)),
+                Tuple.Create(OffsetSpec.ForTimestamp(basetimestamp + 150), new Offset(1)),
+            };
+            
+            foreach (var fixture in testFixtures)
+            {
+                var offsetSpec = fixture.Item1;
+                var offset = fixture.Item2;
+                var topicPartitionOffsetSpecs = new List<TopicPartitionOffsetSpec>
+                {
+                    new TopicPartitionOffsetSpec
+                    {
+                        TopicPartition = new TopicPartition(topic.Name, new Partition(0)),
+                        OffsetSpec = offsetSpec
+                    }
+                };
+                
+                var listOffsetsResult = await adminClient.ListOffsetsAsync(topicPartitionOffsetSpecs,options);
+                foreach(var ListOffsetsResultInfo in listOffsetsResult.ListOffsetsResultInfos)
                 {
                     TopicPartitionOffsetError topicPartition = ListOffsetsResultInfo.TopicPartitionOffsetError;
-                    Assert.Equal((int)topicPartition.Offset, 0);
+                    Assert.Equal(offset, topicPartition.Offset);
                 }
             }
 
-            using (var adminClient = new AdminClientBuilder(new AdminClientConfig { BootstrapServers = bootstrapServers }).Build())
-            {
-                var topicPartitionOffsetSpecs = new List<TopicPartitionOffsetSpec>();
-                topicPartitionOffsetSpecs.Add(new TopicPartitionOffsetSpec {
-                    TopicPartition = new TopicPartition(topic.Name, new Partition(0)),
-                    OffsetSpec = OffsetSpec.Latest()
-                });
-                var ListOffsetsResult = await adminClient.ListOffsetsAsync(topicPartitionOffsetSpecs,options);
-                foreach(var ListOffsetsResultInfo in ListOffsetsResult.ListOffsetsResultInfos)
-                {
-                    TopicPartitionOffsetError topicPartition = ListOffsetsResultInfo.TopicPartitionOffsetError;
-                    Assert.Equal((int)topicPartition.Offset, 3);
-                }
-            }
-
-            using (var adminClient = new AdminClientBuilder(new AdminClientConfig { BootstrapServers = bootstrapServers }).Build())
-            {
-                var topicPartitionOffsetSpecs = new List<TopicPartitionOffsetSpec>();
-                topicPartitionOffsetSpecs.Add(new TopicPartitionOffsetSpec {
-                    TopicPartition = new TopicPartition(topic.Name, new Partition(0)),
-                    OffsetSpec = OffsetSpec.MaxTimestamp()
-                });
-                var ListOffsetsResult = await adminClient.ListOffsetsAsync(topicPartitionOffsetSpecs,options);
-                foreach(var ListOffsetsResultInfo in ListOffsetsResult.ListOffsetsResultInfos)
-                {
-                    TopicPartitionOffsetError topicPartition = ListOffsetsResultInfo.TopicPartitionOffsetError;
-                    Assert.Equal((int)topicPartition.Offset, 1);
-                }
-            }
-
-            topic.Dispose();
-            LogToFile("end   AdminClient_ListOffsets");
-
+            LogToFile("end AdminClient_ListOffsets");
         }
     }
 }